+++ lib/gg_process.dart
@@ -7,4 +7,6 @@
 library;
 
 export 'src/gg_fake_process.dart';
+export 'src/gg_platform_delegate.dart';
+export 'src/gg_process_delegate.dart';
 export 'src/gg_process_wrapper.dart';